# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API AlexTa",
        "description": "Documentation de l'API de l'application AlexTa",
        "contact": {
            "email": "support@alexta.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/commandes": {
            "get": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Afficher toutes les commandes",
                "operationId": "bdaf49272df4c42d1189dcc85f199751",
                "responses": {
                    "200": {
                        "description": "Liste des commandes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Commande"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Créer une nouvelle commande",
                "operationId": "1660514235819d5068a78927af5680f7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status",
                                    "localisation",
                                    "prix_total"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "en attente"
                                    },
                                    "localisation": {
                                        "type": "string",
                                        "example": "Libreville, Akanda"
                                    },
                                    "prix_total": {
                                        "type": "string",
                                        "example": "15000"
                                    },
                                    "service_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "supplement_gabarit_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "supplement_localisation_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Commande créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Commande créée avec succès."
                                        },
                                        "commande": {
                                            "$ref": "#/components/schemas/Commande"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/commandes/{id}": {
            "get": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Afficher une commande spécifique",
                "operationId": "278d5bc14ad859b52a3edfbc0cf7c58e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la commande",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la commande",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Commande"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Commande non trouvée"
                    }
                }
            },
            "put": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Mettre à jour une commande",
                "operationId": "fd9feaee0cfea5a951f23ed6442a73c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la commande à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "livrée"
                                    },
                                    "localisation": {
                                        "type": "string",
                                        "example": "Owendo"
                                    },
                                    "prix_total": {
                                        "type": "string",
                                        "example": "17000"
                                    },
                                    "service_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "supplement_gabarit_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "supplement_localisation_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Commande mise à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Commande mise à jour avec succès."
                                        },
                                        "commande": {
                                            "$ref": "#/components/schemas/Commande"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Commande non trouvée"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Supprimer une commande",
                "operationId": "5ea958f0dab851763a76cff3e8743513",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la commande à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Commande supprimée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Commande supprimée avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Commande non trouvée"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Commande": {
                "title": "Commande",
                "required": [
                    "status",
                    "localisation",
                    "prix_total"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "status": {
                        "type": "string",
                        "example": "en attente"
                    },
                    "localisation": {
                        "type": "string",
                        "example": "Libreville"
                    },
                    "prix_total": {
                        "type": "string",
                        "example": "15000"
                    },
                    "service_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "supplement_gabarit_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "supplement_localisation_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}